Changed contact.html to function properly
This change removes the need for the javascript and research shows that it is widely used for form validation. I used "required" and "pattern" to ensure content was entered correctly. The page was edited to link to a confirmation page. With this change, both contactScript1.js and contactScript2.js can be safely removed.

This page passed validation.