Converted gallery to PHP. Added handler for upload
Converted diaryOverview to PHP so it can pull user content from the database. Created a form handler for the uploadOverlay section. The form handler displays a success or fail message at the conclusion of the upload. CougarApps was inaccessible so haven't been able to test the functionality of the script.